Merge gene interaction and co-expression data

This function merges gene interaction and co-expression data. Function will adjust case sensitivity as needed.

Usage

merge_int_expr(
  gene_int.df,
  gene.coexpr.df,
  int_cols = c("int.col1", "int.col2"),
  coexpr_cols = c("exp.col1", "exp.col2")
)

Arguments

gene_int.df

dataframe with genes interactions

int_cols

names of the two gene names columns in the gene_int.df

gene_coexpr

dataframe with genes co-expression score

expr_cols

names of the two gene names columns in the gene_coexpr.df

Value

A dataframe with gene-gene co-expression score

Examples

merge_int_expr(gene_int.df,gene.coexpr.df, int_cols= c("gene1","gene2"),
 expr_cols=c("Var1","Var2"))
#> Error in merge_int_expr(gene_int.df, gene.coexpr.df, int_cols = c("gene1",     "gene2"), expr_cols = c("Var1", "Var2")): unused argument (expr_cols = c("Var1", "Var2"))