Calculate average rich club coefficient for a list of networks

Calculate average rich club coefficient for a list of networks

Usage

rand_nets_rc(rand_nets, weighted = F)

Arguments

weighted

the option to calculate an edge weighted or unweighted rich-club coefficient

rand_networks

a list of random networks

Value

Average rich club coefficient for a list of networks

Examples

rand_net_clu(rand_networks)
#> Error in rand_net_clu(rand_networks): could not find function "rand_net_clu"